[REFACTOR] #129 - 구조변경
