+++ app/src/main/AndroidManifest.xml
@@ -34,7 +34,7 @@
             android:name="androidx.startup.InitializationProvider"
             android:authorities="${applicationId}.androidx-startup"
             tools:node="remove" />
-        <receiver android:name=".util.WidgetProvider"
+        <receiver android:name=".service.WidgetProvider"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.appwidget.action.APPWIDGET_UPDATE"/>
@@ -43,7 +43,7 @@
                 android:name="android.appwidget.provider"
                 android:resource="@xml/widget_provider"/>
         </receiver>
-        <service android:name=".util.MyRemoteViewsService"
+        <service android:name=".service.ContactRemoteViewsService"
             android:permission="android.permission.BIND_REMOTEVIEWS"/>
     </application>
 </manifest>
