+++ app/src/main/java/com/ivyclub/contact/service/ContactRemoteViewsFactory.kt
@@ -1,10 +1,14 @@
-package com.ivyclub.contact.util
+package com.ivyclub.contact.service
 
 import android.content.Context
 import android.widget.RemoteViews
 import android.widget.RemoteViewsService
 import com.ivyclub.data.model.SimplePlanData
 import com.ivyclub.contact.R
+import com.ivyclub.contact.util.getDayOfMonth
+import com.ivyclub.contact.util.getDayOfWeek
+import com.ivyclub.contact.util.getExactMonth
+import com.ivyclub.contact.util.getExactYear
 import com.ivyclub.data.ContactRepository
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -13,7 +17,7 @@ import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import java.sql.Date
 
-class MyRemoteViewsFactory(
+class ContactRemoteViewsFactory(
     private val context: Context,
     private val repository: ContactRepository
 ) : RemoteViewsService.RemoteViewsFactory {
