+++ app/src/main/res/drawable/widget_background.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <solid
+        android:color="@color/white"/>
+    <corners
+        android:radius="12dp"/>
+
+</shape>
