+++ app/src/main/res/layout/item_widget.xml
@@ -6,7 +6,7 @@
     android:id="@+id/ll_plan"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:layout_marginHorizontal="4dp"
+    android:layout_marginHorizontal="8dp"
     android:layout_marginVertical="4dp"
     android:background="@drawable/bg_plan_list_item"
     android:orientation="vertical"
