+++ app/src/main/res/xml/widget_provider.xml
@@ -5,7 +5,7 @@
     android:initialKeyguardLayout="@layout/widget"
     android:minWidth="300dp"
     android:minHeight="160dp"
-    android:updatePeriodMillis="100000"
+    android:updatePeriodMillis="10000"
     android:resizeMode="horizontal|vertical"
     />
